+++ src/commands/symlink.js
@@ -11,7 +11,7 @@ const _replaceSpaces = require("../functions/replaceSpaces");
 const _fatalError = require("../functions/fatalError");
 
 // Get classes
-const Errors = require("../classes/Errors");
+const { Errors } = require("../classes/Errors");
 const Checks = require("../classes/Checks");
 
 /**