+++ src/variables/constants.js
@@ -16,22 +16,22 @@ const AUTHOR = "Arnav Thorat";
  *
  * Make sure this is a proper version number. Change this with the build (e.g. `78` will make the version `0.7.8`).
  */
-const VERSION = "1.0.0";
+const VERSION = "1.0.3-beta";
 /**
  * The build of the shell.
  *
  * Make sure this is a proper build number. Change this with the version (e.g. `0.7.8` will make the build `78`).
  */
-const BUILD = 100;
+const BUILD = 103;
 
 /**
  * If the shell is in beta or not. `true` if it is in beta, else, `false`.
  */
-const IN_BETA = false;
+const IN_BETA = true;
 /**
  * If the timebomb for the shell has been activated. `true` if the timebomb is activated, else, `false`.
  */
-const TIMEBOMB_ACTIVATED = false;
+const TIMEBOMB_ACTIVATED = true;
 /**
  * The expiry date of the shell. If `TIMEBOMB_ACTIVATED` **and** `IN_BETA` are both `true`, the shell will always check on startup if the current date has surpassed the expiry date.
  *
@@ -51,7 +51,7 @@ const TIMEBOMB_ACTIVATED = false;
  *
  * To see the date in **90** days, [click here](https://www.google.com/search?q=Date+90+days+from+today).
  */
-const EXPIRY_DATE = new Date(2023, 5, 10);
+const EXPIRY_DATE = new Date(2023, 5, 14);
 /**
  * The number of days until the shell will no longer be able to boot. This will not auto-update the `EXPIRY_DATE`.
  */